+++ opencog/atomspaceutils/CMakeLists.txt
@@ -1,7 +1,6 @@
 
 ADD_LIBRARY (atomspaceutils SHARED
 	AtomSpaceUtils
-	FuzzyMatch
 )
 
 ADD_DEPENDENCIES(atomspaceutils opencog_atom_types)
