+++ opencog/atomutils/CMakeLists.txt
@@ -1,7 +1,8 @@
 
 ADD_LIBRARY (atomutils SHARED
-    AtomUtils
-    FindUtils
+	AtomUtils
+	FindUtils
+	FuzzyMatch
 )
 
 ADD_DEPENDENCIES(atomutils opencog_atom_types)
@@ -21,11 +22,12 @@ ELSE (CYGWIN)
 ENDIF (CYGWIN)
 
 INSTALL (FILES
-    AtomUtils.h
-    FindUtils.h
-    FollowLink.h
-    ForeachChaseLink.h
-    HandleMap.h
-    Substitutor.h
-    DESTINATION "include/opencog/atomutils"
+	AtomUtils.h
+	FindUtils.h
+	FollowLink.h
+	ForeachChaseLink.h
+	FuzzyMatch.h
+	HandleMap.h
+	Substitutor.h
+	DESTINATION "include/opencog/atomutils"
 )
