+++ opencog/atomutils/FuzzyMatch.h
@@ -24,6 +24,9 @@
 #ifndef FUZZY_MATCH_H
 #define FUZZY_MATCH_H
 
+#include <opencog/atomspace/Handle.h>
+#include <opencog/atomspace/Link.h>
+
 namespace opencog
 {
 /**
